Refactor plotting services
