+++ m3c2/visualization/__init__.py
@@ -1,1 +1,5 @@
- # Reporting/Plots
+"""Visualization utilities and plotting services."""
+
+from .plot_service import PlotService
+
+__all__ = ["PlotService"]
